+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# HEAD
+
+- Add a chair shape
+- Implement path-tracing algorithm ([PR#10](https://github.com/teozec/image-renderer/pull/10), [#18](https://github.com/teozec/image-renderer/issues/18), [PR#19](https://github.com/teozec/image-renderer/pull/19)).
+- Implement anti-aliasing algorithm ([PR#16](https://github.com/teozec/image-renderer/pull/16)).
+
 # Version 0.2.0
 
 - Implement triangles, boxes and constructive solid geometry ([PR#8](https://github.com/teozec/image-renderer/pull/8)).